move server stuff into its own module
